% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_wilmink.R
\name{daily_lactation}
\alias{daily_lactation}
\title{Calculate daily lactation for a given DIM based on the Wilmink curve}
\usage{
daily_lactation(dim, loga, b, k, d)
}
\arguments{
\item{dim}{days in milk}

\item{loga}{parameter 'a' in the Wilmink curve (log-scale)}

\item{b}{parameter 'b' in the Wilmink curve}

\item{k}{parameter 'k' in the Wilmink curve}

\item{d}{parameter 'd' in the Wilmink curve}
}
\value{
daily lactation value
}
\description{
Calculate daily lactation for a given DIM based on the Wilmink curve
}
