% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_incomplete_model.R, R/woods_curves.R
\name{daily_lactation_f}
\alias{daily_lactation_f}
\title{Daily lactation curve via Wood's curve}
\usage{
daily_lactation_f(n, a, b, c)

daily_lactation_f(n, a, b, c)
}
\arguments{
\item{n}{Days in lactation, i.e. within \code{c(0, 305)}}

\item{a, b, c}{Wood's curve parameters}
}
\value{

}
\description{
Daily lactation curve via Wood's curve
}
