% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entire_herd_scc_curves.R
\docType{data}
\name{entire_herd_scc_curves}
\alias{entire_herd_scc_curves}
\title{This is generated in \code{"notebooks/006_simulation_plan.R"}.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 800 rows and 10 columns.
}
\usage{
entire_herd_scc_curves
}
\description{
The herd does not contain any inflated SCC counts or anything special.
Thus for estimation-procedures the \code{y} needs to be purtubated by some
error, e.g. \code{rnorm(n(), mean = 0, sd = 5)} or similar.
}
\keyword{datasets}
