% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_woods.R
\name{estimate_woods}
\alias{estimate_woods}
\title{Title}
\usage{
estimate_woods(data, logscale = FALSE)
}
\arguments{
\item{data}{}
}
\description{
Title
}
\examples{
# One animal:
data <- simulate_woods(5, 0.01, 0.001)
estimate_woods(data)

# Multiple animals:
library(tidyverse)
data <- lapply(1:5, function(x) simulate_woods(cowID = x)) \%>\% bind_rows()
estimate_woods(data)

FIX: Update model to run one herd, with single cowIDs based on script 013 in notebooks

}
