% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_wilmink.R
\name{f_wilmink}
\alias{f_wilmink}
\title{Wilmink's Curve function}
\usage{
f_wilmink(a, b, d, k, DIM)
}
\arguments{
\item{a, b, c, d}{Parameters for Wilmink's Curve}

\item{DIM}{Days in milk, \verb{[0, 305]} typically and is regarded as the $x$.}
}
\value{
\code{log(SCC)}
}
\description{
Wilmink's Curve function
}
