% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mastitis_shape.R
\name{find_mastititis_shapes}
\alias{find_mastititis_shapes}
\title{Title}
\usage{
find_mastititis_shapes(par = c(1, 1))
}
\arguments{
\item{par}{}
}
\value{

}
\description{
Title
}
\note{
The \code{mode} in the output should be the same as the target
mode provided.
}
