% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lactation_phases.R
\name{lactation_phases}
\alias{lactation_phases}
\title{Title}
\usage{
lactation_phases(peak_location)
}
\arguments{
\item{peak_location}{In Wood's Model, that is \code{b/c}, but it could be something else?}
}
\value{
Breaks corresponding to the phase that the lactation period is currently in.
}
\description{
Title
}
