% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mastitis_shape_parameters.R
\docType{data}
\name{mastitis_shape_parameters}
\alias{mastitis_shape_parameters}
\title{These are the parameters found using
\code{"notebooks/05_mastitis_shape_find.R"}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 2 columns.
}
\usage{
mastitis_shape_parameters
}
\description{
and they represent the most proper shape of extra SCC cells added due
to a mastitis infection.
}
\keyword{datasets}
