% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_parameters.R
\name{sample_parameters}
\alias{sample_parameters}
\title{Let us define a "parameter sampler"}
\usage{
sample_parameters(n = 1)
}
\arguments{
\item{n}{\code{a} must be non-negative
\code{d} must be non-negative}
}
\description{
Let us define a "parameter sampler"
}
