% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_scc_curves.R
\name{sample_scc_curves}
\alias{sample_scc_curves}
\title{Title}
\usage{
sample_scc_curves(parameters_df, herd_size)
}
\arguments{
\item{variables}{}
}
\value{

}
\description{
Title
}
\examples{
Generate individual cow parameters from the first lactation:

  #'
  herd_df \%>\%
    pivot_longer(cols = c(cow_a, cow_b, cow_c)) \%>\%

    identity() \%>\% {
      ggplot(.) +
        aes(value, group = interaction(name)) +
        geom_density() +
        geom_rug(
          data = . \%>\%
            group_by(name, lactation, herd) \%>\%
            slice(1) \%>\%
            pivot_longer(cols = c(herd_a, herd_b, herd_c),
                         names_to = "herd_name",
                         values_to = "herd_value"),
          aes(herd_value)) +
        facet_wrap(~name, ncol = 1, scales = "free") +
        NULL
    }
  #'

herd_full_df <- herd_df \%>\%
  expand_grid(
    DIM = seq.default(5, 305)
  ) \%>\%
  mutate(daily_lactation_f  =
           pmap_dbl(select(., n = DIM, a = cow_a, b = cow_b, c = cow_c),
                    daily_lactation_f))
#'
herd_level_curve_df <- herd_df \%>\%
  slice(1) \%>\%
  select(-cow_a, -cow_b, -cow_c, -cow_id) \%>\%
  expand_grid(
    DIM = seq.default(5, 305)
  ) \%>\%
  mutate(daily_lactation_f  =
           pmap_dbl(select(., n = DIM, a = herd_a, b = herd_b, c = herd_c),
                    daily_lactation_f))
#'
#'
herd_full_df \%>\%

  identity() \%>\% {
    ggplot(.) +
      aes(DIM, daily_lactation_f) +
      geom_line(aes(color = cow_id,
                    group = interaction(cow_id)),
                alpha = .5,
                show.legend = FALSE) +

      geom_line(aes(color = "herd-level"),
                data = herd_level_curve_df,
                linetype = "dashed",
                size = 2,
                show.legend = FALSE) +
      scale_color_viridis_d() +
      NULL
  }
}
