% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_extend_inflation_scc.R
\name{shift_extend_inflation_scc}
\alias{shift_extend_inflation_scc}
\title{Shift & Extend a value}
\usage{
shift_extend_inflation_scc(x, shift_by, total_length = 305, fill = 0)
}
\arguments{
\item{shift_by}{Non-negative integer representing the number of entries
to shift to the right.}

\item{total_length}{Length of the resulting value.}
}
\value{
Returns numeric that is \code{total_length} long with values
from \code{x} padded with \code{fill}
}
\description{
Shift & Extend a value
}
\examples{

}
