% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate SCC data for dairy cows}
\usage{
simulate_data(n_cows, a_mean, a_sd, b_mean, b_sd, c_mean, c_sd)
}
\arguments{
\item{n_cows}{Number of cows to simulate}

\item{a_mean}{Mean value for parameter 'a'}

\item{a_sd}{Standard deviation for parameter 'a'}

\item{b_mean}{Mean value for parameter 'b'}

\item{b_sd}{Standard deviation for parameter 'b'}

\item{c_mean}{Mean value for parameter 'c'}

\item{c_sd}{Standard deviation for parameter 'c'}
}
\value{
A data frame with simulated SCC data for cows
}
\description{
Simulate SCC data for dairy cows
}
\examples{
simulate_data(10, 5, 0.1, 2, 0.2, 3, 0.3)
}
