% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_woods.R
\name{simulate_woods}
\alias{simulate_woods}
\title{Simulate data from a woods curve}
\usage{
simulate_woods(
  a = 5,
  b = 0.01,
  c = 0.001,
  dim = seq(30, 300, by = 30),
  cowID = "CKR_1"
)
}
\arguments{
\item{a}{desc}

\item{b}{desc}

\item{c}{desc}

\item{dim}{which dim do we want?}
}
\value{
a series of simulated data from 1 cow with 30 dim between observations
}
\description{
Simulate data from a woods curve
}
\examples{
data <- simulate_woods()

}
