% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woods_curves.R
\name{total_lactation}
\alias{total_lactation}
\title{Title}
\source{
From \href{https://www.nature.com/articles/216164a0.pdf}{Algebraic Model of the Lactation Curve in Cattle}
}
\usage{
total_lactation(a, b, c)
}
\arguments{
\item{a, b, c}{}
}
\value{

}
\description{
Title
}
\examples{

}
