% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mastitis_shape.R,
%   R/mastitis_scc_inflated.R
\name{mastitis_shape}
\alias{mastitis_shape}
\title{Title}
\usage{
mastitis_shape(
  severity_scale,
  n = 100,
  x = NULL,
  mastitis_shape_par = mastitis_shape_par
)

mastitis_shape(
  severity_scale,
  n = 100,
  x = NULL,
  mastitis_shape_par = mastitis_shape_par
)
}
\arguments{
\item{severity_scale}{Between \code{c(0, 1)}}

\item{n}{Number of points between \code{c(0, 21)} returned}

\item{x}{If specified, the returned points are based on this \code{x}.}
}
\value{



}
\description{
Title

Title
}
