% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_wilmink.R
\name{simulate_wilmink}
\alias{simulate_wilmink}
\title{Simulate data from a wilmink curve}
\usage{
simulate_wilmink(herd, animals, interval, a, b, k, d, resid)
}
\arguments{
\item{a}{desc}

\item{b}{desc}

\item{k}{desc}

\item{d}{desc}

\item{dim}{which DIM do we want?}
}
\value{
simulated data on herd level with an interval of 28 DIM between osbervations
}
\description{
Simulate data from a wilmink curve
}
\examples{
data <- simulate_wilmink()

}
